TooltipSettings are now chart local
