+++ src/ducks/Studio/Chart/Fullscreen/index.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useRef } from 'react'
 import { linearScale, logScale } from '@santiment-network/chart/scales'
 import Header from './Header'
 import ChartCanvas from '../Canvas'
+import { useWidgetMetricLabeling } from '../../Widget/utils'
 import { useTimeseries } from '../../timeseries/hooks'
 import { extractMirrorMetricsDomainGroups } from '../../../Chart/utils'
 import { MirroredMetric } from '../../../dataHub/metrics/mirrored'
@@ -34,7 +35,8 @@ const FullscreenChart = ({
   selectedLineState,
   isDrawingState,
   isNewDrawingState,
-  setIsICOPriceDisabled
+  setIsICOPriceDisabled,
+  cursorType
 }) => {
   const [settings, setSettings] = useState(studioSettings)
   const [options, setOptions] = useState(studioOptions)
@@ -82,6 +84,8 @@ const FullscreenChart = ({
     },
     [metrics]
   )
+
+  useWidgetMetricLabeling(chartRef, metrics, settings)
 
   function changeTimePeriod (fromDate, toDate) {
     const interval = getNewInterval(fromDate, toDate)
@@ -149,6 +153,7 @@ const FullscreenChart = ({
           isDomainGroupingActive ? domainGroups : mirrorDomainGroups
         }
         selectedLineState={selectedLineState}
+        cursorType={cursorType}
         isDrawingState={isDrawingState}
         isNewDrawingState={isNewDrawingState}
         onPointHover={undefined}
